creacion del componente logo y se le hace append en las vistas de home, login y signup
